% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coexp_arma}
\alias{coexp_arma}
\title{cal_coexp
This function calculates the coexpression patterns between genes
and returns the coexpression matrix.}
\usage{
coexp_arma(X)
}
\arguments{
\item{X}{Input binarized cell (row) by gene (column) matrix}
}
\value{
Coexpression matrix
}
\description{
cal_coexp
This function calculates the coexpression patterns between genes
and returns the coexpression matrix.
}
\author{
Qi Gao
}
