% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcellsgroup.R
\name{getcellsgroup}
\alias{getcellsgroup}
\title{getcellsgroup}
\usage{
getcellsgroup(
  so,
  index_gene_up,
  index_gene_down,
  score,
  estimated.nonfeatured.proportion
)
}
\arguments{
\item{so}{Seurat Object}

\item{index_gene_up}{a char vector of up-regulated markers (as the rownames of seurat object)}

\item{index_gene_down}{a char vector of down-regulated markers (as the rownames of seurat object)}

\item{score}{cellular score to use GSVA/CFS}

\item{estimated.nonfeatured.proportion}{estimated featured cells proportion}
}
\value{
Updated Seurat Object with new cell group in the metadata
}
\description{
Regroup cells in the Seurat object
}
\details{
This function divides cells into two groups according to their scores
calculated by gene markers' expression levels. The score threshold is set
based on the estimation of featured cells proportion in the population.
}
