% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBLightning.R
\name{runBLightning}
\alias{runBLightning}
\title{runBLightning}
\usage{
runBLightning(
  so,
  upregulated.markers,
  downregulated.markers,
  score = "CFS",
  estimated.nonfeatured.proportion = 0.9,
  connectivity.cutoff = ceiling(0.3 * (length(upregulated.markers) +
    length(downregulated.markers))),
  num.variablefeatures = 2000,
  alpha.genes = 0.05,
  max.iter = 10000
)
}
\arguments{
\item{so}{Preprocessed Seurat Object}

\item{upregulated.markers}{a char vector of established up-regulated gene markers (as the rownames of seurat object)}

\item{downregulated.markers}{a char vector of established down-regulated gene markers (as the rownames of seurat object)}

\item{score}{the type of cellular score used to group cells: CFS/GSVA}

\item{estimated.nonfeatured.proportion}{estimated proportion of nonfeatured cells
in the whole cell population, default is 0.9}

\item{connectivity.cutoff}{number of established genes that candidates have to be correlated with, default is the smallest integer that is greater than or equal to the number of input genes}

\item{num.variablefeatures}{number of highly variable genes used to calculate the null distribution of connectivity, default is 2000}

\item{alpha.genes}{TypeI error of identifying differentally expressed genes, default is 0.05}

\item{max.iter}{Maximum iterations to run, default is 10000.}
}
\value{
A list that contains the number of iterations, new upregulated gene markers,
new downregulated gene markers,input upregulated gene markers, input downregulated gene markers,  pvalues of new upregulated gene markers,
pvaluds of new downregualted markers
}
\description{
An iterative fishing method for gene markers
}
\details{
This function is the main function of B-Lightning
}
