% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getscore.R
\name{getscore}
\alias{getscore}
\title{Calculate scores of cells}
\usage{
getscore(score = "CFS1", index_gene_up, index_gene_down, subcounts)
}
\arguments{
\item{index_gene_up}{names of upregulated markers}

\item{index_gene_down}{names of downregulated markers}

\item{subcounts}{count matrix (gene markers by cells)
#' @return score vector}
}
\description{
This function calculates scores of each cell based on gene markers' expression levels.
}
